[compiler] Fix clustering rules resulting in TensorRT offloads with boolean shape tensor inputs
TensorRT does not allow boolean shape tensors as inputs. A long-standing
issues is that we currently don't have a good way of dealing with this
constraint in our clustering algorithm.

The crux of the issue is that we clustering for different backends sequentially
instead of at the same time/jointly. This means that it's difficult to reason
about the effect of simply perturbing the boundary of a cluster or inserting
cast operations to change the boundary type -- we need a way to ensure that the
inserted casts are offloaded to some backend and we want to avoid having to
create such extra ops in the first place.

This commit fixes the problem temporoarily by disallowing operations which
likely have boolean shape tensor operands from being clustered to the
TensorRT backend at all. Generally this is OK since such operations
can almost always be offloaded to the host.

GitOrigin-RevId: ecb492dd0674d9f73dac7c2999bc0a72019c4857