Add tensor methods for cast, copy, and shape manipulation ops (#688)
Currently all operations are implemented as free functions, giving
consistency to the API. To improve the user experience, readability, and
transfer from other frameworks, we are adding a subset of operations to
also be available as tensor methods. Currently, we are including
operations that don't modify the underlying data, instead changing
attributes like the device, datatype, or view. These are frequently
"chained" operations in frameworks like torch, improving the transfer
experience and readability.