[cmake] Add option to use relative file paths in logging and debug info
This is useful for a) builds in production where we don't what full
paths to be used and b) to make logging more readable. Changing the
option will generally invalidate your compiler cache for the project,
though, so try to avoid turning it on and off. Only works if compiler
supports '-ffile-prefix-map'.

GitOrigin-RevId: 386eba1dd7749c9ac045fee5cfed0eb24c1b0be7