[executor/Runtime] Require RuntimeClient for RuntimeSession construction
We previously implicitly required a RuntimeClient for the RuntimeSession
to work properly, but the API was enforcing it in a very strange way (
when RuntimeSession was changed to require RuntimeClient when a function
returned results, some effort was made to minimize API change).

This change fixes the API to be more consistent. RuntimeSession must be
constructed from a RuntimeClient. It is now unnecessary to explicitly
manage the lifetime of RuntimeSession objects relative to the RuntimeClient
in Python since RuntimeSession retains a reference to the RuntimeClient
on the C++ side.

In addition, we add an interface abstract method to `RuntimeSession`
to enforce enforce consistency in backend implementations like
LuaRuntimeSession.

GitOrigin-RevId: ce694914dd50f9eb0e4fd150b580dff634e48b2f