[integrations/python] Fix DLPack synchronization behavior
The previous behavior of `MemRefValue.__dlpack__` ignored the 'stream',
'dl_device', and 'copy' arguments of the `__dlpack__` method
specification. After recent runtime changes, this exposed intermittent
bugs (e.g. in cupy tests) due to incorrect synchronization This change
fixes the behavior so that we explicitly return an error in situations
that we don't handle and properly handle the 'stream' operand to ensure
correct synchronization.

GitOrigin-RevId: 5a2f600f28084b34aface4eea24e8dea58b95a62