[executor] Introduce a pass for lowering aggregates at function boundaries
This change introduces a pass for lowering aggregates (e.g.
`!executor.table`) at function boundaries. This is in preparation for
merging the new compiler-runtime inteface mechanism which requires us
to explicitly handle aggregate arguments and results.

There are three possible lowering modes:

1. `direct`: Lower aggregates at function boundaries by passing
    aggregates by-value. If this option is specified, the pass
    has no effect; the IR is left unchanged.

2. `indirect`: Lower aggregates at function boundaries by passing
    aggregates through indirect buffers. At call sites, we use
    stack-allocated buffers (alloca) to hold the aggregate.
    Aggregate results are promoted to destination pointer
    operands and the caller must allocate that memory.

3. `unpacked`: Lower aggregates at function boundaries by passing
    them as a flat list of scalars. This expands all aggregates into
    many scalar arguments or results.

If `indirect` or `unpacked` mode is specified, then the pass
also inserts `executor.result` arguments to specify the original
result index of the function prior to modification.

GitOrigin-RevId: f88269062ec20545f84a61b8a5c2450083ecd519