[compiler] Improve handling of "result slot" tracking
This change is a step towards implementing the new compiler-runtime
interface specification. Previously, we used a UnitAttr named
"plan.result_arg" to specify that an argument is the promoted
DPS equivalent of a result. However, this obviously lacks information
concerning "which result" it is equivalent to (with respect to the
results of the original high-level function given as input to the
compiler). The fact that our current compiler-runtime interface works
is a side-effect of the fact that we don't fully support modes
where a compiled function could return some results through DPS
and other results through normal returns -- the runtime currently has
no way of distinguishing how the user's original function signature
maps down to the arguments/results of the compiled function.

This change just replaces 'plan|executor.result_arg' with
'plan|executor.result_slot' and enforces that an integer attribute
must be given as the value. This then allows for checking certain
assumptions (e.g. that all the DPS arguments are strictly increasing
in terms of numbering).

We also add a new pass called 'verify-input-and-assign-slots' which
is run early in the compiler pipeline and ensures that the results of
the original function are assigned unique slot numbers. These numbers
are then propagated throughout transformations like the bufferization
pipeline and Executor aggregate lowering.

The runtime executable schema is not changed, so this change only
affects the compiler side of the interface.

GitOrigin-RevId: a2b16ebdb9ea351f1530259f14dc9f32b4db6594