[tensorrt] Check if CUDA device is available before translating to TensorRT.
This MR adds a check to ensure that a CUDA device is available before translating to TensorRT
begins. Right now, first step when translation begins is createing TensorRT builder context, which
fails if no CUDA device is available. Error we get is failed to create TensorRT builder context but
error message is not explicit saying that CUDA device is not available.
This MR alleviates this issue by printing a more explicit error message.

GitOrigin-RevId: 6b6ee457ca6bde2560579b62af8e1615bad38016