[mlir-tensorrt] fix the way that layer names are set for TensorRT (#702)
This PR fixes the way that layer names are set for TensorRT.

~~Previously there was a bug in the `getUniqueName` function that caused
it to generate names like `name_0_1_2_3_4_5_6_7` instead of `name_7`.~~
This was not a problem upstream.

Also fixed the way that call sites are translated to layer names.
Previous if a `Location` tracked the entire stack trace (function `a`
calls `b` calls `c` calls `d` ), then the name would be set as
`a:LINE_NUMBER`. Now, the layer name is set to `b:LINE -> c:LINE ->
d:LINE` to provide enough information to identify where an layer comes
from.

Co-authored-by: Matthew Francis-Landau <matthew.francis-landau@gm.com>