[compiler] Add pass for converting `tensorrt` dialect IO bounds attributes to `plan` dialect
Certain users use `tensorrt.(shape_profile|value_bounds)` attributes
to specify argument shape/value bounds at the frontend, but the compilation
pipeline mostly expects Plan dialect attributes outside of the TensorRT dialect.
In the `stablehlo-to-executable` pipeline, we were previously doing attribute
conversion from `tensorrt->plan` in the `plan-materialize-shape-calculations`
pass, which was a bit adhoc. Furthermore, when we introduce the ABI V1 stubs
into the pipeline, we will need to convert these attributes prior to creation of
the ABI functions. This commit adds a dedicated pass for converting these
attributes and hoists the transformation up to the top of the
stablehlo-to-executable pipeline.

GitOrigin-RevId: 66413e940204b6da0f55ca0f0335630bd77db049