[common] Update `mtrt::StatusOr` implementation
Updates `mtrt::StatusOr` to use `std::variant`. Improves the API and
adds ref-qualified access variants for better optimization when
`StatusOr<T>` is an r-value. Fixes a couple minor issues in status
check macros used in the runtime. Improves the API for better
consistency.

The `mtrt::Status` and `mtrt::StatusOr` classes are now completely
defined in the header files, which makes it possible to use them in
e.g. plugin DSOs that should not link against `libMTRT`.

GitOrigin-RevId: 96c85ad22147552c51805b6ceebe70d985cf53c4