[executor] Fix buffer deallocation edge case
Fixes an issue where we didn't properly model when memrefs
returned from ABI wrapper functions are "live out", resulting
in possibly trying to deallocate a memref whose ownership is supposed
to be passed to the caller.

The fix requires that `exeuctor.abi.send` pass through its
argument value, allowing it to be returned from the ABI wrapper
function. This lets us to properly model that results are live-out
without having to define a dedicated `executor.abi.func` operation.

As a consequence, we have to propgate this change through the passes
that analyze or modify `exuector.abi.send` operations, but they are
fairly mechanical.

GitOrigin-RevId: 0114949b22ecc64afa20e122e55d46e66a0bc732