[compiler] Fix bug in `plan-buffer-results-to-out-params` pass
This change fixes a bug where the `plan-buffer-results-to-out-params` pass
would consider a returned memref value as 'hoistable' by moving operations
producing that value to the call site and passing the memref through
arguments. The logic was not checking the case where the cluster of
operations to be hoisted has multiple users of different cluster output
values and cannot be replaced with a single argument.

GitOrigin-RevId: 85247618d9d13ef86b4b671c32c16e6d8fdf1d7d