[mlir-tensorrt][cmake] Refactor compilation options into dedicated CMake module, add extra checks mode
This change consolidates handling of global compilation options (that are
specific to MTRT and not handled as part of `HandleLLVMOptions`) into
a dedicated CMake module. It adds additional CMake cache options for
enabling '-Werror' globally and for addition additional warnings and
runtime checks recommended by
https://best.openssf.org/Compiler-Hardening-Guides/Compiler-Options-Hardening-Guide-for-C-and-C++.html.

By default, `MLIR_TRT_ENABLE_EXTRA_CHECKS` and `MLIR_TRT_ENABLE_WERROR` are
set to `OFF`, but the idea is to enable them in CI in the future.

In addition, fix some lingering warnings and build issues discovered
when testing with the extra checks mode and Werror enabled:
- Supress deprecated declaration warnings related to TensorRT
- Fix some warnings when NCCL is enabled.
- Fix test flags when ORTools is OFF but Shardy is ON.

GitOrigin-RevId: 39f5e91697a20a65da2a5374b90366aba005f889