[compiler] Fix handling of replacements in ToLoopsOpInterface
This change fixes an issue where the implementations of the
ToLoopsOpInterface for different operations were using different policies
for replacement inside of the `OpType::toLoops` interface
implementations. This could result in a crash or incorrect output IR.
This change fixes the issue -- `OpType::toLoops` should not perform
a replacement itself; rather the replacement should be made by the
caller.

GitOrigin-RevId: 7b5b901cc03bef17b49024df9a56df08f292e99f