[compiler] Improve `plan-create-closed-regions` interaction with CompilerBackendAttrInterface
Previously, only clusters marked with the `plan.tensorrt_backend`
actually interacted with the `plan-create-closed-regions` pass. The
other cluster types were marked as "not requiring closure". This was
very confusing and resulted in a set of code related to cluster
closure and outlining that was never actually invoked. Ideally, all
cluster types proceed uniformly through the same set of stages
(cluster->closure->outline) so that it's easier to reason about how
each cluster kind is transformed.

To resolve this, CompilerBackendAttrInterface was updated to include
methods for detailing whether DPS calling convention is supported
and/or preferred the given `plan.inline_group` operation instance. If
DPS is not supported, then `plan.inline_closed_alloc_group` is used
during the closure step. Resolves some issues in `mlir::outlineRegionOp`
since it previously assumed that the cluster region op had no block
arguments.

In addition, the backends now all declare a function for determining
whether a producer should be cloned into the region or passed by argument
during the region closure step. This reduces the amount of duplicate
code present for that logic during the plan segmentation pipeline.

Finally, the cluster-to-region-op formation logic now correctly orders
all yielded values deterministically based on an ordering of the uses,
obviating the need for complicated hooks for reordering the yielded
values in the TensorRTToExecutable pass.

GitOrigin-RevId: 4a179d1f329fd8089d082a019b3691d040ed4d32