[compiler] Add `mtrt-scf-strength-reduction` pass (#737)
Adds a pass for normalizing `scf.while` loops that can be transformed
to `scf.for` loops with floating-point induction variables. The pass
walks over all `scf.while` loops and looks at each loop carried
argument.
It identifies arguments carried by the while loop that are float scalars
which are constant-initialized, compared against a constant limit, and
have
a constant add/sub updatestep within the body. It identifies a scale
factor using the `step` if necessary to confirm the arithmetic can be
converted to integer arithmetic. It then rewrites the loop to use an
integer, converting to an `scf.for` loop if possible.

GitOrigin-RevId: e675dc71fb13cffa2b6188426916d31f6e274526