[mlir-tensorrt] Integrate Kernel backend
This change integrates the 'kernel' subproject, which provides a
kernel generation pipeline which builds off of upstream MLIR dialects
and the LLVM NVPTX backend.

The main purpose is to enable generation of kernels for StableHLO
operations that cannot be offloaded to the main TensorRT backend due
to functional limitations or performance considerations.

The `kernel` subproject can be built in a standalone fashion, but
anything which requires actually running a kernel (e.g. integration
tests) is done through the main `compiler` project.

The input abstraction for pipelines declared under `kernel` is Linalg
IR or one of the high-level operations declared as part of the Kernel
dialect. Examples of high-level operations include `kernel.sort` and
`kernel.scatter`, which are specifically designed to correspond with
StableHLO operations for the best possible coverage.

The main compiler declares a backend 'KernelBackend' which is used to
specify which operations should/can be offloaded to the kernel
generation pipeline. While this specification happens on StableHLO,
after the initial cluster+outline step amongst all backends, the
functions targeted toward kernel generation are processed by inlining
back into the main entrypoint function(s), converting to Linalg or
Kernel dialect, optimizing on Linalg if possible, and then
re-clustering again to outline to separate kernels. This is done to
get better fusion results.

Once the code targeted to kernel generation is isolated in Linalg or
Kernel IR form, one of two things happen. For some operations like
`kernel.sort`, we may directly emit a SIMT kernel in MLIR using
standard MLIR dialects. Otherwise, for Linalg input, we generate a
kernel via a gradual lowering/tiling process utilizing MLIR's
Transform IR infrastructure.

In any case, the generation logic produces SIMT kernels, which are
`func.func` nested within `gpu.module` operations. These kernels are
then bufferized and lowered to NVVM. The LLVM NVPTX backend is
invoked, and the PTX is embedded in the module as a global. JIT
compilation to cuBin occurs in the runtime.

We do a joint bufferization across the entire host program and
generated kernels after they have been lowered to SIMT form. This is
enabled through the "multi-module bufferization" transformation
defined by the Plan dialect. All generated kernels are bufferized such
that they assume fully dynamic layouts. Then, after the whole program
is bufferized, we inspect all callers in the host program and try to
greedily refine layouts of arguments, which can often result in
simplifiation to identity layouts.

The Kernel dialect provides two different operations for invoking
kernels from the host program: `kernel.call` and `kernel.ext_call`.
The first is a DPS-style operation and can only be used with kernels
whose signatures follow a defined format which adheres to MLIR's DPS
convention. This is used for kernels generated from linalg. The second
is for non-DPS style kernels where the signature could take any form.
The caller must be sure to correctly capture the effects associated
with each operand. This version is used for kernels that are directly
emitted since it is more flexible. Both call operations can be
bufferized.

Linalg input problems are pattern-matched against specific patterns
determined by the "TransformSchedule" classes defined in the kernel
dialect. Currently these include ScatterSchedule and FallbackSchedule.
The ScatterSchedule is a more specialized schedule that is used to
generate SIMT kernels for the `kernel.scatter` operation. The
FallbackSchedule is the most general and should be able to handle most
Linalg operations. It is not meant to be a performant generator -- we
just want to generate functional SIMT kernels to avoid compilation
failures for StableHLO when we can't offload to TensorRT. These
schedules provide parameter (e.g. tile shapes) estimation and a
function to generate specific Transform-IR for each kernel. This
Transform-IR is then interpreted to yield the SIMT kernel.

Co-authored-by: Zixin Huang <zixinh@nvidia.com>
Co-authored-by: Bo Wang <bowa@nvidia.com>
Co-authored-by: Sagar Shelke <sagshelke@nvidia.com>

GitOrigin-RevId: 0d6c0ddd5107eb7414579e61003ff05fd08860c8